+++ MANIFEST.in
@@ -1,1 +1,1 @@
-include app/trumps/*
+include gbtrump/trumps/*
