+++ entry_points.cfg
@@ -1,2 +1,2 @@
 [console_scripts]
-gbtrump = app:main
+gbtrump = gbtrump:main
