+++ gbtrump/logger.py
