+++ setup.cfg
@@ -12,5 +12,5 @@ license_file = LICENSE
 [options]
 include_package_data = True
 packages=find:
-scripts = app/__main__.py
+scripts = gbtrump/__main__.py
 entry_points = file: entry_points.cfg
